app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
streamlit
sentence-transformers
pillow
requests
numpy
pandas
